src/test/java/io/github/jendker/model/projection/GroupReadModelTest.java
package io.github.jendker.model.projection;

import io.github.jendker.model.Task;
import io.github.jendker.model.TaskGroup;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.Set;

import static org.assertj.core.api.Assertions.assertThat;

class GroupReadModelTest {

    @Test
    @DisplayName("Should crate null deadline for group when no task deadlines")
    void constructor_noDeadlines_createsNullDeadline(){
        //given
        var source = new TaskGroup();
        source.setDescription("foo");
        source.setTasks(Set.of(new Task("bar", null)));

        //when
        var result = new GroupReadModel((source));

        //then
        assertThat(result).hasFieldOrPropertyWithValue("deadline", null);

    }

}